% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDeID.R
\name{fileDeID}
\alias{fileDeID}
\title{De-identify machine extraction data}
\usage{
fileDeID(
  filetodeid,
  fd_varname_mrn = "PAT_MRN",
  variablestoremove = character(0),
  variablestoblank = character(0),
  datevariablestodateshift = character(0),
  dateformat = "\%Y-\%m-\%d",
  datetimevariablestodateshift = character(0),
  datetimeformat = "\%Y-\%m-\%d \%H:\%M:\%OS",
  xwalk,
  compare_mrn_numeric = attr(xwalk, "compare_mrn_numeric"),
  outputfile = NULL,
  verbose = 0
)
}
\arguments{
\item{filetodeid}{A single csv file to de-identify.}

\item{fd_varname_mrn}{Name of mrn variable in file to de-identify. Default is PAT_MRN.}

\item{variablestoremove}{Names of variables in extraction file to remove. Default is character(0), no variables to remove.}

\item{variablestoblank}{Names of variables in extraction file to blank. The variable will remain in the output file but will be "" for all rows. Default is character(0), no variables to blank.}

\item{datevariablestodateshift}{Names of variables that are dates and should be date-shifted. Default is character(0), no variables to shift.}

\item{dateformat}{Format of date variables in the extraction file.  Default is "\%Y-\%m-\%d" corresponding to 4-digit year, hyphen, 2-digit month, hyphen, 2-digit day.}

\item{datetimevariablestodateshift}{Names of variables that are datetimes and should be date-shifted. Default is character(0), no variables to shift.}

\item{datetimeformat}{Format of date variables in the extraction file.  Default is "\%Y-\%m-\%d \%H:\%M:\%OS" corresponding to 4-digit year, hyphen, 2-digit month, hyphen, 2-digit day, space, 2-digit (24) hour, colon, 2-digit minute, colon 2-digit second.}

\item{xwalk}{data.frame containing cross walk information. Usually the output from loadxwalks().}

\item{compare_mrn_numeric}{Should MRNs be compared as numeric variables? Usually this is a good idea because leading 0s may have been dropped during processing. Default is whatever was used to create xwalk, which is TRUE by default.}

\item{outputfile}{Name of file to write de-identified data. If NULL (the default) the data are not written, only returned from the function.}

\item{verbose}{Higher values produce more output to console.  Default is 0, no output.}
}
\value{
data.frame with variables tokenized, dateshifted, removed, and blanked, as requested.
}
\description{
De-identify machine extraction data
}
\examples{
# ddd <- fileDeID(
#   filetodeid = "J:/FORUM Data/CSV/2005/UM_OPH_VISUAL_FIELD.csv",
#   fd_varname_mrn = "MRN",
#   variablestoremove = c("X", "MRN0", "DOB", "Age", "Sex", "Institution"),
#   variablestoblank = "Name",
#   datevariablestodateshift = "TestDate",
#   dateformat = "\%Y\%m\%d",
#   datetimevariablestodateshift = "TestDateTime",
#   xwalk = xwalk,
#   verbose = 2
# )
}
