% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadxwalks.R
\name{loadxwalks}
\alias{loadxwalks}
\title{Load crosswalk(s) needed to de-identify machine data. One crosswalk between mrns and tokenized mrns. Optionally, a second file with mrns and date shift variable}
\usage{
loadxwalks(
  tokenfile = NULL,
  t_varname_mrn = "PAT_MRN",
  t_varname_mrn_token = "PAT_MRN_T",
  dateshiftfile = NULL,
  ds_varname_mrn = "PAT_MRN",
  ds_varname_dateshift = "SHIFT_NUM",
  compare_mrn_numeric = TRUE,
  verbose = 0
)
}
\arguments{
\item{tokenfile}{Name of the mrn-tokenized mrn crosswalk file.}

\item{t_varname_mrn}{Name of the mrn variable in the token file.  Default is "PAT_MRN".}

\item{t_varname_mrn_token}{Name of the tokenized mrn variable in the token file.  Default is "PAT_MRN_T".}

\item{dateshiftfile}{Name of the mrn-date shift file.}

\item{ds_varname_mrn}{Name of the mrn variable in the date shift file.  Default is "PAT_MRN".}

\item{ds_varname_dateshift}{Name of the date shift variable. Default is "SHIFT_NUM"}

\item{compare_mrn_numeric}{Should MRNs be compared as numeric variables? Usually this is a good idea because leading 0s may have been dropped during processing. Default is TRUE.}

\item{verbose}{Higher values produce more output to console.  Default is 0, no output.}
}
\value{
A data.frame with columns PAT_MRN, PAT_MRN_T, and SHIFT_NUM if dateshiftfile is not NULL.
}
\description{
Load crosswalk(s) needed to de-identify machine data. One crosswalk between mrns and tokenized mrns. Optionally, a second file with mrns and date shift variable
}
\examples{
dataloc <- system.file("extdata", package = "DeIDmachinedata")
fn1 <- sprintf("\%s/xwalk1.csv", dataloc)
fn2 <- sprintf("\%s/xwalk2.csv", dataloc)
xwalk <- loadxwalks(
  tokenfile = fn1,
  t_varname_mrn = "PAT_MRN",
  t_varname_mrn_token = "PAT_MRN_T",
  dateshiftfile = fn2,
  ds_varname_mrn = "PAT_MRN",
  ds_varname_dateshift = "SHIFT_NUM",
  compare_mrn_numeric = TRUE,
  verbose = 2)
xwalk
}
