% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDeID.R
\name{fileDeID}
\alias{fileDeID}
\title{De-identify machine extraction data}
\usage{
fileDeID(
  filetodeid,
  fd_varname_mrn = "PAT_MRN",
  variablestoremove = character(0),
  variablestoblank = character(0),
  datevariablestodateshift = character(0),
  dateformat = "\%Y-\%m-\%d",
  datetimevariablestodateshift = character(0),
  datetimeformat = "\%Y-\%m-\%d \%H:\%M:\%OS",
  epochvariablestodateshift = character(0),
  separator = ",",
  separator_out = ",",
  xwalk,
  compare_mrn_numeric = attr(xwalk, "compare_mrn_numeric"),
  outputfile = NULL,
  usefread = TRUE,
  verbose = 0L
)
}
\arguments{
\item{filetodeid}{character. Filename of a single flat file to de-identify.}

\item{fd_varname_mrn}{character. Name of MRN variable in file to de-identify. Default is "PAT_MRN".}

\item{variablestoremove}{character vector. Names of variables in extraction file to remove. Default is \code{character(0)}, no variables to remove.}

\item{variablestoblank}{character vector. Names of variables in extraction file to blank. The variable will remain in the output file but will be '' for all rows. Default is \code{character(0)}, no variables to blank.}

\item{datevariablestodateshift}{character vector. Names of variables that are dates and should be date-shifted. Default is \code{character(0)}, no variables to shift.}

\item{dateformat}{character. Format of date variables in the extraction file.  Default is "\%Y-\%m-\%d" corresponding to 4-digit year, hyphen, 2-digit month, hyphen, 2-digit day.}

\item{datetimevariablestodateshift}{character vector. Names of variables that are datetimes and should be date-shifted. Default is \code{character(0)}, no variables to shift.}

\item{datetimeformat}{character. Format of datetime variables in the extraction file.  Default is "\%Y-\%m-\%d \%H:\%M:\%OS" corresponding to 4-digit year, hyphen, 2-digit month, hyphen, 2-digit day, space, 2-digit (24) hour, colon, 2-digit minute, colon 2-digit second.}

\item{epochvariablestodateshift}{character vector. Names of variables that represent
Unix epoch time (seconds since 1970-01-01 00:00:00 UTC) and should be shifted.
Values may be stored as integer seconds or floating-point seconds (allowing
fractional seconds). Default is \code{character(0)}, no variables to shift.}

\item{separator}{character. Field separator in filetodeid (input file). Default is \code{","}.}

\item{separator_out}{character. Field separator to use in outputfile. Default is \code{","}.}

\item{xwalk}{data.frame containing crosswalk information. Usually the output from \code{\link[=loadxwalks]{loadxwalks()}}.}

\item{compare_mrn_numeric}{logical. Should MRNs be compared as numeric variables? Usually this is a good idea because leading 0s may have been dropped during processing. Default is whatever was used to create xwalk, which is \code{TRUE} by default.}

\item{outputfile}{character. Name of file to write de-identified data. If \code{NULL} (the default) the data are not written, but only returned from the function. An additional option is the special value \code{"SOURCE_"}, which causes the output to be written to the same filename as the input but prepended with "SOURCE_".}

\item{usefread}{If \code{TRUE} (default), use \code{data.table::fread()} and \code{data.table::fwrite()}. If \code{FALSE}, use \code{utils::read.csv()} and \code{utils::write.csv()}. \code{TRUE} is usually preferable as \code{FALSE} results in double quotes around almost all values when producing the output file.}

\item{verbose}{integer. Higher values produce more output to console.  Default is 0, no output.}
}
\value{
(invisibly) data.frame (even if \code{usefread == TRUE}) with variables tokenized, date-shifted, removed, and/or blanked, as requested.
}
\description{
This function is used to de-identify a flat file using a crosswalk (with
specific structure, see \code{loadxwalks()}).
De-identification options include replacing a patient MRN by a tokenized MRN
(using the crosswalk), removing a column, 'blanking' a column
(the column remains in the dataset but all values are ""), and shifting a
date, datetime, or Unix epoch time variable (using the crosswalk).
}
\details{
Date variables are shifted by adding \code{SHIFT_NUM} days.
Datetime variables are shifted by adding \code{86400 * SHIFT_NUM} seconds.
Epoch variables are shifted by adding \code{86400 * SHIFT_NUM} seconds directly
to the numeric epoch value.

All epoch values are assumed to represent seconds since
1970-01-01 00:00:00 UTC. No timezone conversion is performed.
}
\examples{
dataloc <- system.file("extdata", package = "DeIDmachinedata")
fn1 <- sprintf("\%s/xwalk1.csv", dataloc)
fn2 <- sprintf("\%s/xwalk2.csv", dataloc)
xwalk <- loadxwalks(tokenfile = fn1, dateshiftfile = fn2)
fn3 <- sprintf("\%s/pentacam_UCH.csv", dataloc)
deidfile <- fileDeID(
  filetodeid = fn3,
  fd_varname_mrn = "Pat-ID:",
  variablestoremove = c("Last Name:", "First Name:", "D.o.Birth:"),
  variablestoblank = "Exam Comment:",
  datevariablestodateshift = "Exam Date:",
  dateformat = "\%m/\%d/\%Y",
  xwalk = xwalk,
  outputfile = NULL,
  verbose = 2)
deidfile
}
