% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDeID.R
\name{fileDeID}
\alias{fileDeID}
\title{De-identify machine extraction data}
\usage{
fileDeID(
  filetodeid,
  fd_varname_mrn = "PAT_MRN",
  variablestoremove = character(0),
  variablestoblank = character(0),
  datevariablestodateshift = character(0),
  dateformat = "\%Y-\%m-\%d",
  datetimevariablestodateshift = character(0),
  datetimeformat = "\%Y-\%m-\%d \%H:\%M:\%OS",
  separator = ",",
  separator_out = ",",
  xwalk,
  compare_mrn_numeric = attr(xwalk, "compare_mrn_numeric"),
  outputfile = NULL,
  usefread = TRUE,
  verbose = 0
)
}
\arguments{
\item{filetodeid}{A single csv file to de-identify.}

\item{fd_varname_mrn}{Name of mrn variable in file to de-identify. Default is PAT_MRN.}

\item{variablestoremove}{Names of variables in extraction file to remove. Default is character(0), no variables to remove.}

\item{variablestoblank}{Names of variables in extraction file to blank. The variable will remain in the output file but will be "" for all rows. Default is character(0), no variables to blank.}

\item{datevariablestodateshift}{Names of variables that are dates and should be date-shifted. Default is character(0), no variables to shift.}

\item{dateformat}{Format of date variables in the extraction file.  Default is "\%Y-\%m-\%d" corresponding to 4-digit year, hyphen, 2-digit month, hyphen, 2-digit day.}

\item{datetimevariablestodateshift}{Names of variables that are datetimes and should be date-shifted. Default is character(0), no variables to shift.}

\item{datetimeformat}{Format of date variables in the extraction file.  Default is "\%Y-\%m-\%d \%H:\%M:\%OS" corresponding to 4-digit year, hyphen, 2-digit month, hyphen, 2-digit day, space, 2-digit (24) hour, colon, 2-digit minute, colon 2-digit second.}

\item{separator}{Field separator in filetodeid. Default is ",".}

\item{separator_out}{Field separator to use in outputfile. Default is ",".}

\item{xwalk}{data.frame containing cross walk information. Usually the output from loadxwalks().}

\item{compare_mrn_numeric}{Should MRNs be compared as numeric variables? Usually this is a good idea because leading 0s may have been dropped during processing. Default is whatever was used to create xwalk, which is TRUE by default.}

\item{outputfile}{Name of file to write de-identified data. If NULL (the default) the data are not written, but only returned from the function. An additional option is the special value "SOURCE_", which causes the output to be written to the same filename as the input but prepended with "SOURCE_".}

\item{usefread}{If TRUE (default), use data.table::fread and data.table::fwrite. If FALSE, use utils::read.csv and utils::write.csv. TRUE is preferable as the latter adds double quotes around almost all values when producing output file.}

\item{verbose}{Higher values produce more output to console.  Default is 0, no output.}
}
\value{
data.frame (invisibly and even if usefread == TRUE) with variables tokenized, dateshifted, removed, and blanked, as requested.
}
\description{
De-identify machine extraction data
}
\examples{
dataloc <- system.file("extdata", package = "DeIDmachinedata")
fn1 <- sprintf("\%s/xwalk1.csv", dataloc)
fn2 <- sprintf("\%s/xwalk2.csv", dataloc)
xwalk <- loadxwalks(tokenfile = fn1, dateshiftfile = fn2)
fn3 <- sprintf("\%s/pentacam_UCH.csv", dataloc)
deidfile <- fileDeID(
  filetodeid = fn3,
  fd_varname_mrn = "Pat-ID:",
  variablestoremove = c("Last Name:", "First Name:", "D.o.Birth:"),
  variablestoblank = "Exam Comment:",
  datevariablestodateshift = "Exam Date:",
  dateformat = "\%m/\%d/\%Y",
  xwalk = xwalk,
  outputfile = NULL,
  verbose = 2)
deidfile # Note last test was on a person not in the crosswalk
}
